2-6

+++ src/com/alpha/PracticalWork25/work4/Runner.java
@@ -1,5 +1,7 @@
 package com.alpha.PracticalWork25.work4;
 
+import java.math.BigDecimal;
+
 public class Runner {
 
     public void run() {
@@ -16,11 +18,11 @@ public class Runner {
         System.out.println("i1 == i4 : " + (i1 == i4));
         System.out.println("i1 equals i4 : " + (i1.equals(i4)));
 
-        System.out.println("\nSum = " + sum(1, 3.3f));
-        System.out.println("Sum = " + sum(Integer.valueOf("1"), Float.valueOf("3.3")));
+        System.out.println("\nSum = " + sum(1, 3.3d));
+        System.out.println("Sum = " + sum(Integer.valueOf("1"), Double.valueOf("3.3")));
     }
 
-    private Double sum (Integer i, Float f) {
-        return Double.valueOf(i + f);
+    private double sum (Integer i, Double f) {
+        return i + f;
     }
 }
